refactor(services): split session_detector_test.go by scenario group
CodeScene flagged session_detector_test.go at Code Health 5.12/10 and
2645+ LOC, by far the largest file in its flagged batch. Investigation
confirmed the low score is a size/organization artifact, not a logic
smell: the file was a flat sequence of independent, well-isolated
table-driven tests with no shared complexity, just accreted scenario
groups that were never split out (unlike compact/hold, dead-process,
no-substantive-activity, and workflow, which already got their own
sibling files in earlier PRs).

Split the remaining content into four new files along the same
scenario-group convention, matching the shape of the production code
under test:

- session_detector_activity_test.go: working/waiting/ready
  classification transitions, plus the IsAgentDone/NeedsUserAttention
  unit tests.
- session_detector_lifecycle_test.go: session removal, process-exit
  handling, /continue re-creation, seedFromDisk (incl. the consent
  gate), and the PreToolUse permission hook.
- session_detector_pid_test.go: PID assignment and the /clear
  same-PID reuse cleanup path.
- session_detector_subagent_test.go: parent-child state propagation
  (the largest remaining group, explicitly called out in the issue).

Every test function body was moved byte-for-byte (verified via a
sorted line-set diff against the pre-split file) — this is a pure
reorganization, no behavior or assertion changes. `goimports` pruned
each new file's import block. All 60 test functions are preserved
with no drops or duplicates.

`go test ./core/... -race -count=1` is green (verified with reduced
build parallelism — the shared dev machine was under heavy unrelated
load during this session, which intermittently flakes several
fixed-sleep tests in this package on both the split and pre-split
tree alike; unrelated to this change).

Closes #809
